pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spectralclean"
version = "0.1.0"
description = "Spectral eigenvector-based noise detection and cleaning for COCO object detection datasets."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Alan Peng"},
]
keywords = [
    "object-detection",
    "dataset-cleaning",
    "spectral-analysis",
    "label-noise",
    "COCO",
    "computer-vision",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
dependencies = [
    "numpy>=1.24",
    "opencv-python>=4.8",
    "Pillow>=9.0",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "torch>=2.0",
    "transformers>=4.36",
    "tqdm>=4.65",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[project.scripts]
spectralclean = "spectralclean.cli:main"

[project.urls]
Homepage = "https://github.com/alanpeng/SpectralClean"
Repository = "https://github.com/alanpeng/SpectralClean"

[tool.setuptools.packages.find]
include = ["spectralclean*"]
